requirements.txt
yt-dlp>=2024.12.0
google-api-python-client==2.110.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
apscheduler==3.10.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
google-generativeai>=0.8.0
gTTS==2.5.0
flask==3.0.0
PyJWT>=2.8.0
# GPU 가속을 위한 패키지
faster-whisper>=1.1.0
torch>=2.0.0
torchaudio>=2.0.0
# CUDA 11.8 또는 12.1 (RTX 4060 Ti 호환)
# pip install torch torchaudio --index-url https://download.pytorch.org/whl/cu118

# macOS Apple Silicon (M1/M2/M3) - 맥에서만 설치
# pip install mlx-whisper
